Move saving logic to background page
And implement react- chrome-redux
